Small Changes from presentation feedback

+++ flutter-final-01/map_my_shopping_v1/lib/ui/search_page.dart
@@ -70,7 +70,7 @@ class SearchPageState extends State<SearchPage> {
       String name = "$term item number + $i";
       String desc = "example of additional search results";
       double price = rnd.nextDouble();
-      String aisle = thingy[faker.randomGenerator.integer(26)] +
+      String aisle = thingy[faker.randomGenerator.integer(9)] +
           faker.randomGenerator.integer(30).toString();
       String dep = deps[rnd.nextInt(deps.length)];
       results.add(ShopItem(name, desc, price, aisle, dep));
